1.0.1

+++ main.cpp
@@ -4,11 +4,17 @@
 #include <chrono>
 #include <thread>
 #include <cstdlib>
+#include <Windows.h>
+#include <filesystem>
+
+#define _CRT_SECURE_NO_DEPRECATE
+#define _CRT_SECURE_NO_WARNINGS
+#define _CRT_NONSTDC_NO_DEPRECATE
+#pragma warning(disable : 4996)
 
 int wingetInstall() {
     std::cout << "WARNING! IF YOU GET AN ERROR, TRY TO INSTALL WITH ANOTHER METHOD! \n";
     std::string userInput1;
-
     std::cout << "Continue? (Y/N): ";
     std::cin >> userInput1;
 
@@ -35,14 +41,17 @@ int wingetInstall() {
             outputFile.close();
 
             std::cout << "Success" << std::endl;
-        } else {
+        }
+        else {
             std::cout << "Failed to edit PowerShell Profile" << std::endl;
         }
-    } else if (userInput1 == "N" || userInput1 == "No" || userInput1 == "n") {
+    }
+    else if (userInput1 == "N" || userInput1 == "No" || userInput1 == "n") {
         std::cout << "Exiting..." << std::endl;
         std::this_thread::sleep_for(std::chrono::milliseconds(800));
         exit(0);
-    } else {
+    }
+    else {
         std::cout << "Invalid input. Please enter Y or N." << std::endl;
     }
     return 0;
@@ -76,20 +85,44 @@ int manualInstall() {
             outputFile.close();
 
             std::cout << "Success" << std::endl;
-        } else {
+        }
+        else {
             std::cout << "Failed to edit PowerShell Profile" << std::endl;
         }
-    } else if (userInput2 == "N" || userInput2 == "No" || userInput2 == "n") {
+    }
+    else if (userInput2 == "N" || userInput2 == "No" || userInput2 == "n") {
         std::cout << "Exiting..." << std::endl;
         std::this_thread::sleep_for(std::chrono::milliseconds(800));
         exit;
-    } else {
+    }
+    else {
         std::cout << "Invalid input. Please enter Y or N." << std::endl;
     }
     return 0;
 }
 
+std::string getUsername() {
+    char* username;
+    size_t len;
+    errno_t err = _dupenv_s(&username, &len, "USERPROFILE");
+    return std::string(username);
+}
+
+
+bool doesDirectoryExist(const std::string& path) {
+    std::filesystem::path p(path);
+    return std::filesystem::exists(p) && std::filesystem::is_directory(p);
+}
+
 int main() {
+    std::string username = getUsername();
+    std::string path = "C:\\Users\\" + username + "\\AppData\\Local\\Programs\\oh-my-posh\\bin";
+
+    if (doesDirectoryExist(path)) {
+        std::cout << "Oh-My-Posh is installed." << std::endl;
+        exit;
+    }
+
     std::string userInput;
     std::cout << "Do you want to install using WinGet? (Y/N): ";
     std::cin >> userInput;
@@ -97,11 +130,13 @@ int main() {
     if (userInput == "Y" || userInput == "Yes" || userInput == "y") {
         system("cls");
         wingetInstall();
-    } else if (userInput == "N" || userInput == "No" || userInput == "n") {
+    }
+    else if (userInput == "N" || userInput == "No" || userInput == "n") {
         std::cout << "Installing with other method..." << std::endl;
         system("cls");
         manualInstall();
-    } else {
+    }
+    else {
         std::cout << "Invalid input. Please enter Y or N." << std::endl;
     }
